requirements.txt
build>=0.6.0
twine>=3.4.2
scikit-learn==1.0.2
pandas==1.4.2
grpio>=1.54.0
grpcio-tools>=1.54.0pp
pymongo>=4.3.3
tensorflow>=2.12.0
